Add new first section about cube complexes concerned with metric
Another section was neede, which is concerned with the metric
preliminaries for CAT(0) cube complexes. We need to define CAT(0)
spaces in full generality and embed cube complexes in the broader
setting of M_kappa-polyhedral complexes. This will be achieved with
this section.

Right now the definition of a CAT(kappa) space and length spaces has
been given.