Change definition of convex polyhedral complex to conf/ cubing
I changed the definition to be in accordance with a cubing. Hence all
cubes are always embedded. This makes the objects combinatorially
easier to handle and is no restriction in the case of CAT(0) cube
complexes, because they are always cubings. In general one should be
able to go over to the first barycentric subdivision to achieve our
definition. I also added remarks to explain everything in this regard.